Caption coverage: a linter that names what it could not check, and a timeline
A caption track's faults exist only at runtime, in one place, for a moment. An
overlap puts two lines on screen at once. A cue that outlives the audio hangs over
silence. A stretch with nothing in it is invisible in a list of numbers and obvious
as a hole. Reading the table cannot show you any of that, and scrubbing the whole
clip to find it is why nobody checks.

LintCaptionTrack reports overlaps, out-of-order cues, zero-length cues, empty text,
cues too fast to read, cues past the end of the clip, and long gaps -- each with the
numbers in the message, because "there is an overlap" does not tell an author
whether it is a rounding error or a real mistake.

The property that matters more than any individual rule: when the clip length is
unknown, the two checks that need it are returned in `notRun`, BY NAME, and the
panel prints them ABOVE the findings. A linter that quietly skips half its rules
and then says "no problems found" converts an unchecked track into a checked one in
the author's head, which is worse than having no linter. For the same reason
Coverage() answers -1 rather than 0 without a clip length: 0.0 would read as "none
of it is captioned", a specific and wrong claim.

Nor is the clip length inferred from the last cue's end. That would make coverage
read near 100% on every track including one that stops captioning halfway through,
which is precisely the case the number exists to catch.

Coverage is the UNION of the cue windows, not their sum. Counting an overlap twice
could report 150%, which reads as a broken tool rather than as the overlap -- and
the overlap already has its own finding.

Gaps are checked at the start and at the end too, not only between cues. A linter
that only looked between them would miss a silent opening minute and a silent
closing one, which are the two places captions most often stop.

The Caption Track panel draws the cues against time: gaps shaded, cues alternating
rows so an overlap reads as two stacked blocks, findings marked in place, a
click-to-scrub playhead, and clicking a finding moves the playhead to it. Its empty
state distinguishes "no data assets loaded" from "assets loaded, none is a caption
track", because those have different fixes.

This needed EditorPanel widened from u32 to u64. It was FULL at 1u << 31, and three
panels had already been pushed out of the mask into standalone bools whose
visibility is not saved with the rest -- two of them say so in their own comments
("no EditorPanel bits left", "only bit 31 of that mask is still free").
EditorSettings::visiblePanels went to u64 with it; left at u32 it would have
silently dropped every panel above bit 31 on save, so a new panel would be visible
until the next restart and then never again.

21 lint tests. 185/185 pass.

Not yet verified visually: the panel is a hand-drawn surface and a screenshot is
the only way to judge its layout, but the machine is running a fullscreen game and
capturing would have meant taking focus from it. The drawing code follows the three
recorded ImGui rules (panel-phase only, sRGB-to-linear before IM_COL32, explicit
sizes for every AddText since FontGlobalScale does not reach it).

+++ Engine/include/Enjin/Editor/EditorSettings.h
@@ -195,7 +195,11 @@ struct EditorSettings {
     void AddRecentVisualScriptNode(const std::string& nodeTypeId);
 
     // Layout persistence
-    u32 visiblePanels = 319;          // EditorPanel bitmask - shipped default panel set
+    // EditorPanel bitmask - shipped default panel set. u64 because the enum is:
+    // stored as u32 it would have silently dropped every panel above bit 31 on
+    // save, so a newly added panel would be visible until the next restart and
+    // then never again.
+    u64 visiblePanels = 319;
     f32 leftPanelWidth = 0.18f;
     f32 rightPanelWidth = 0.25f;
     f32 bottomPanelHeight = 0.22f;
