+++ Engine/src/Editor/EditorSettings.cpp
@@ -615,7 +615,7 @@ bool EditorSettings::Load(const std::string& path) {
         }
 
         // Layout persistence
-        if (j.contains("visiblePanels")) visiblePanels = j["visiblePanels"].get<u32>();
+        if (j.contains("visiblePanels")) visiblePanels = j["visiblePanels"].get<u64>();
         if (j.contains("leftPanelWidth")) leftPanelWidth = std::clamp(j["leftPanelWidth"].get<f32>(), 0.05f, 0.5f);
         if (j.contains("rightPanelWidth")) rightPanelWidth = std::clamp(j["rightPanelWidth"].get<f32>(), 0.05f, 0.5f);
         if (j.contains("bottomPanelHeight")) bottomPanelHeight = std::clamp(j["bottomPanelHeight"].get<f32>(), 0.05f, 0.5f);
